Commonize RoborazziOptions and processOutputImageAndReport in core
