+++ include-build/roborazzi-core/src/commonMain/kotlin/com/github/takahirom/roborazzi/RoborazziRecordFilePathStrategy.kt
@@ -1,0 +1,16 @@
+package com.github.takahirom.roborazzi
+
+@ExperimentalRoborazziApi
+sealed interface RoborazziRecordFilePathStrategy {
+  val propertyValue: String
+
+  object RelativePathFromCurrentDirectory : RoborazziRecordFilePathStrategy {
+    override val propertyValue: String
+      get() = "relativePathFromCurrentDirectory"
+  }
+
+  object RelativePathFromRoborazziContextOutputDirectory : RoborazziRecordFilePathStrategy {
+    override val propertyValue: String
+      get() = "relativePathFromRoborazziContextOutputDirectory"
+  }
+}
