+++ include-build/roborazzi-core/src/iosMain/kotlin/com/github/takahirom/roborazzi/CanScreenshot.kt
@@ -1,0 +1,3 @@
+package com.github.takahirom.roborazzi
+
+actual fun canScreenshot(): Boolean = true
