+++ include-build/roborazzi-core/src/iosMain/kotlin/com/github/takahirom/roborazzi/DefaultCaptureType.kt
@@ -1,0 +1,5 @@
+package com.github.takahirom.roborazzi
+
+actual fun defaultCaptureType(): RoborazziOptions.CaptureType {
+  throw NotImplementedError("defaultCaptureType(would be dump) is not implemented on iOS")
+}
